Update to Yichun1024?

+++ MOFServer/MOFServer/Handler/MOFCharacter.cs
@@ -670,11 +670,11 @@ public class MOFCharacter
 public class Transactor
 {
     //在交易的位置與物品
-    public Dictionary<int, Item> Items =new Dictionary<int, Item>();
+    public Dictionary<int, Item> Items = new Dictionary<int, Item>();
 
     //存一下放進交易欄的東西原本從哪來
     public Dictionary<int, Item> BackItem = new Dictionary<int, Item>();
-    public int Rubi = 0;
+    public long Rubi = 0;
     public bool IsReady = false;
 
 }
